+++ packages/twenty-server/src/engine/metadata-modules/dashboard-sync/dashboard.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { DashboardSyncService } from 'src/engine/metadata-modules/dashboard/services/dashboard-sync.service';
+import { DashboardSyncService } from 'src/engine/metadata-modules/dashboard-sync/services/dashboard-sync.service';
 import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
 import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
 
@@ -9,4 +9,4 @@ import { TwentyORMModule } from 'src/engine/twenty-orm/twenty-orm.module';
   providers: [DashboardSyncService],
   exports: [DashboardSyncService],
 })
-export class DashboardModule {}
+export class DashboardSyncModule {}
