+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/page-layout-widget.module.ts
@@ -4,7 +4,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ApplicationModule } from 'src/engine/core-modules/application/application.module';
 import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
 import { WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.entity';
-import { DashboardModule } from 'src/engine/metadata-modules/dashboard/dashboard.module';
+import { DashboardSyncModule } from 'src/engine/metadata-modules/dashboard-sync/dashboard.module';
 import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
 import { FlatPageLayoutWidgetModule } from 'src/engine/metadata-modules/flat-page-layout-widget/flat-page-layout-widget.module';
 import { PageLayoutWidgetController } from 'src/engine/metadata-modules/page-layout-widget/controllers/page-layout-widget.controller';
@@ -28,7 +28,7 @@ import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspa
     WorkspaceManyOrAllFlatEntityMapsCacheModule,
     FlatPageLayoutWidgetModule,
     ApplicationModule,
-    DashboardModule,
+    DashboardSyncModule,
   ],
   controllers: [PageLayoutWidgetController],
   providers: [
