+++ packages/twenty-server/src/engine/metadata-modules/page-layout/page-layout.module.ts
@@ -5,7 +5,7 @@ import { ApplicationModule } from 'src/engine/core-modules/application/applicati
 import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
 import { I18nModule } from 'src/engine/core-modules/i18n/i18n.module';
 import { WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.entity';
-import { DashboardModule } from 'src/engine/metadata-modules/dashboard/dashboard.module';
+import { DashboardSyncModule } from 'src/engine/metadata-modules/dashboard-sync/dashboard.module';
 import { WorkspaceManyOrAllFlatEntityMapsCacheModule } from 'src/engine/metadata-modules/flat-entity/services/workspace-many-or-all-flat-entity-maps-cache.module';
 import { FlatPageLayoutTabModule } from 'src/engine/metadata-modules/flat-page-layout-tab/flat-page-layout-tab.module';
 import { FlatPageLayoutWidgetModule } from 'src/engine/metadata-modules/flat-page-layout-widget/flat-page-layout-widget.module';
@@ -40,7 +40,7 @@ import { WorkspaceMigrationV2Module } from 'src/engine/workspace-manager/workspa
     ApplicationModule,
     PageLayoutTabModule,
     PageLayoutWidgetModule,
-    DashboardModule,
+    DashboardSyncModule,
   ],
   controllers: [PageLayoutController],
   providers: [
